+++ apps/jaspr_pad/lib/components/playground/output/execution_iframe.dart
@@ -1,31 +1,30 @@
 import 'package:jaspr/components.dart';
 import 'package:jaspr_riverpod/jaspr_riverpod.dart';
 
+import '../../utils/node_reader.dart';
 import 'execution_service.dart';
 import 'execution_service.imports.dart';
 
 class ExecutionIFrame extends StatelessComponent {
-  const ExecutionIFrame({Key? key}) : super(key: key);
+  const ExecutionIFrame({super.key});
 
   @override
   Iterable<Component> build(BuildContext context) sync* {
-    yield FindChildNode(
-      onNodeAttached: (node) {
-        if (kIsWeb) {
-          var iframe = context.read(iframeProvider);
-          if (iframe == null || iframe != node.nativeElement) {
-            context.read(iframeProvider.notifier).state = node.nativeElement as IFrameElementOrStubbed;
-          }
+    yield DomNodeReader(
+      onNode: (node) {
+        var iframe = context.read(iframeProvider);
+        if (iframe == null || iframe != node) {
+          context.read(iframeProvider.notifier).state = node as IFrameElementOrStubbed;
         }
       },
-      child: DomComponent(
-        tag: 'iframe',
+      child: iframe(
         id: 'frame',
+        src: 'https://dartpad.dev/scripts/frame_dark.html',
+        sandbox: 'allow-scripts allow-popups',
         attributes: {
-          'sandbox': 'allow-scripts allow-popups',
           'flex': '',
-          'src': 'https://dartpad.dev/scripts/frame_dark.html',
         },
+        [],
       ),
     );
   }