+++ apps/jaspr_pad/lib/components/playground/panels/console_panel.dart
@@ -4,23 +4,20 @@ import 'package:jaspr_riverpod/jaspr_riverpod.dart';
 import '../output/execution_service.dart';
 
 class ConsolePanel extends StatelessComponent {
-  const ConsolePanel({Key? key}) : super(key: key);
+  const ConsolePanel({super.key});
 
   @override
   Iterable<Component> build(BuildContext context) sync* {
     var messages = context.watch(consoleMessagesProvider);
 
-    yield DomComponent(
-      tag: 'div',
+    yield div(
       styles: Styles.flexbox(direction: FlexDirection.column),
       classes: ['console', 'custom-scrollbar'],
-      children: [
+      [
         for (var msg in messages)
-          DomComponent(
-            tag: 'span',
-            styles: Styles.box(width: Unit.zero),
-            child: Text(msg, rawHtml: true),
-          ),
+          span(styles: Styles.box(width: Unit.zero), [
+            text(msg, rawHtml: true),
+          ]),
       ],
     );
   }