+++ apps/jaspr_pad/lib/components/playground/panels/issues_panel.dart
@@ -6,47 +6,39 @@ import '../../../providers/edit_provider.dart';
 import '../../../providers/issues_provider.dart';
 
 class IssuesPanel extends StatelessComponent {
-  const IssuesPanel({Key? key}) : super(key: key);
+  const IssuesPanel({super.key});
 
   @override
   Iterable<Component> build(BuildContext context) sync* {
     var issues = context.watch(issuesProvider);
 
-    yield DomComponent(
-      tag: 'div',
-      styles: Styles.flexbox(direction: FlexDirection.column),
-      children: [
-        for (var issue in issues) IssueItem(issue),
-      ],
-    );
+    yield div(styles: Styles.flexbox(direction: FlexDirection.column), [
+      for (var issue in issues) IssueItem(issue),
+    ]);
   }
 }
 
 class IssueItem extends StatelessComponent {
-  const IssueItem(this.issue, {Key? key}) : super(key: key);
+  const IssueItem(this.issue, {super.key});
 
   final Issue issue;
 
   @override
   Iterable<Component> build(BuildContext context) sync* {
-    yield DomComponent(
-      tag: 'span',
-      classes: ['issue-item', issue.kind.name],
-      events: {
-        'click': (e) {
-          context.read(fileSelectionProvider(issue.sourceName).notifier).state = issue.location;
-          context.read(activeDocIndexProvider.notifier).state =
-              context.read(fileNamesProvider).indexOf(issue.sourceName);
-        }
-      },
-      children: [
-        DomComponent(
-          tag: 'i',
-          classes: ['material-icons'],
-          child: Text(issue.kind.name),
-        ),
-        Text(issue.message),
-      ],
-    );
+    yield span(classes: [
+      'issue-item',
+      issue.kind.name
+    ], events: {
+      'click': (e) {
+        context.read(fileSelectionProvider(issue.sourceName).notifier).state = issue.location;
+        context.read(activeDocIndexProvider.notifier).state = context.read(fileNamesProvider).indexOf(issue.sourceName);
+      }
+    }, [
+      i(
+        classes: ['material-icons'],
+        [text(issue.kind.name)],
+      ),
+      text(issue.message),
+    ]);
   }
 }