+++ apps/jaspr_pad/lib/models/sample.mapper.dart
@@ -1,7 +1,8 @@
 // coverage:ignore-file
 // GENERATED CODE - DO NOT MODIFY BY HAND
 // ignore_for_file: type=lint
-// ignore_for_file: unused_element
+// ignore_for_file: unused_element, unnecessary_cast
+// ignore_for_file: strict_raw_type, inference_failure_on_untyped_parameter
 
 part of 'sample.dart';
 
@@ -16,11 +17,6 @@ class SampleMapper extends ClassMapperBase<Sample> {
     return _instance!;
   }
 
-  static T _guard<T>(T Function(MapperContainer) fn) {
-    ensureInitialized();
-    return fn(MapperContainer.globals);
-  }
-
   @override
   final String id = 'Sample';
 
@@ -32,7 +28,7 @@ class SampleMapper extends ClassMapperBase<Sample> {
   static const Field<Sample, int> _f$index = Field('index', _$index);
 
   @override
-  final Map<Symbol, Field<Sample, dynamic>> fields = const {
+  final MappableFields<Sample> fields = const {
     #id: _f$id,
     #description: _f$description,
     #index: _f$index,
@@ -46,38 +42,38 @@ class SampleMapper extends ClassMapperBase<Sample> {
   final Function instantiate = _instantiate;
 
   static Sample fromMap(Map<String, dynamic> map) {
-    return _guard((c) => c.fromMap<Sample>(map));
+    return ensureInitialized().decodeMap<Sample>(map);
   }
 
   static Sample fromJson(String json) {
-    return _guard((c) => c.fromJson<Sample>(json));
+    return ensureInitialized().decodeJson<Sample>(json);
   }
 }
 
 mixin SampleMappable {
   String toJson() {
-    return SampleMapper._guard((c) => c.toJson(this as Sample));
+    return SampleMapper.ensureInitialized().encodeJson<Sample>(this as Sample);
   }
 
   Map<String, dynamic> toMap() {
-    return SampleMapper._guard((c) => c.toMap(this as Sample));
+    return SampleMapper.ensureInitialized().encodeMap<Sample>(this as Sample);
   }
 
   SampleCopyWith<Sample, Sample, Sample> get copyWith => _SampleCopyWithImpl(this as Sample, $identity, $identity);
   @override
   String toString() {
-    return SampleMapper._guard((c) => c.asString(this));
+    return SampleMapper.ensureInitialized().stringifyValue(this as Sample);
   }
 
   @override
   bool operator ==(Object other) {
     return identical(this, other) ||
-        (runtimeType == other.runtimeType && SampleMapper._guard((c) => c.isEqual(this, other)));
+        (runtimeType == other.runtimeType && SampleMapper.ensureInitialized().isValueEqual(this as Sample, other));
   }
 
   @override
   int get hashCode {
-    return SampleMapper._guard((c) => c.hash(this));
+    return SampleMapper.ensureInitialized().hashValue(this as Sample);
   }
 }
 
@@ -122,11 +118,6 @@ class SampleResponseMapper extends ClassMapperBase<SampleResponse> {
     return _instance!;
   }
 
-  static T _guard<T>(T Function(MapperContainer) fn) {
-    ensureInitialized();
-    return fn(MapperContainer.globals);
-  }
-
   @override
   final String id = 'SampleResponse';
 
@@ -136,7 +127,7 @@ class SampleResponseMapper extends ClassMapperBase<SampleResponse> {
   static const Field<SampleResponse, String> _f$error = Field('error', _$error);
 
   @override
-  final Map<Symbol, Field<SampleResponse, dynamic>> fields = const {
+  final MappableFields<SampleResponse> fields = const {
     #project: _f$project,
     #error: _f$error,
   };
@@ -149,39 +140,40 @@ class SampleResponseMapper extends ClassMapperBase<SampleResponse> {
   final Function instantiate = _instantiate;
 
   static SampleResponse fromMap(Map<String, dynamic> map) {
-    return _guard((c) => c.fromMap<SampleResponse>(map));
+    return ensureInitialized().decodeMap<SampleResponse>(map);
   }
 
   static SampleResponse fromJson(String json) {
-    return _guard((c) => c.fromJson<SampleResponse>(json));
+    return ensureInitialized().decodeJson<SampleResponse>(json);
   }
 }
 
 mixin SampleResponseMappable {
   String toJson() {
-    return SampleResponseMapper._guard((c) => c.toJson(this as SampleResponse));
+    return SampleResponseMapper.ensureInitialized().encodeJson<SampleResponse>(this as SampleResponse);
   }
 
   Map<String, dynamic> toMap() {
-    return SampleResponseMapper._guard((c) => c.toMap(this as SampleResponse));
+    return SampleResponseMapper.ensureInitialized().encodeMap<SampleResponse>(this as SampleResponse);
   }
 
   SampleResponseCopyWith<SampleResponse, SampleResponse, SampleResponse> get copyWith =>
       _SampleResponseCopyWithImpl(this as SampleResponse, $identity, $identity);
   @override
   String toString() {
-    return SampleResponseMapper._guard((c) => c.asString(this));
+    return SampleResponseMapper.ensureInitialized().stringifyValue(this as SampleResponse);
   }
 
   @override
   bool operator ==(Object other) {
     return identical(this, other) ||
-        (runtimeType == other.runtimeType && SampleResponseMapper._guard((c) => c.isEqual(this, other)));
+        (runtimeType == other.runtimeType &&
+            SampleResponseMapper.ensureInitialized().isValueEqual(this as SampleResponse, other));
   }
 
   @override
   int get hashCode {
-    return SampleResponseMapper._guard((c) => c.hash(this));
+    return SampleResponseMapper.ensureInitialized().hashValue(this as SampleResponse);
   }
 }