+++ packages/jaspr/lib/browser.dart
@@ -1,7 +1,7 @@
 library browser;
 
-export 'jaspr.dart' hide runApp, AppBinding, NativeDomNode;
+export 'jaspr.dart' hide runApp, AppBinding;
 export 'src/browser/browser_binding.dart';
 export 'src/browser/clients.dart';
-export 'src/browser/native_interop.dart';
+export 'src/browser/dom_render_object.dart';
 export 'src/browser/run_app.dart';