+++ packages/jaspr/lib/jaspr.dart
@@ -3,7 +3,6 @@ export 'package:meta/meta.dart'
 
 export 'src/components/async.dart';
 export 'src/components/basic.dart';
-export 'src/components/child_node.dart';
 export 'src/components/style.dart';
 export 'src/foundation/annotations.dart';
 export 'src/foundation/basic_types.dart';
@@ -15,9 +14,6 @@ export 'src/foundation/scheduler.dart';
 export 'src/foundation/sync.dart';
 export 'src/foundation/synchronous_future.dart';
 export 'src/framework/framework.dart';
-export 'src/stub/interop_stub.dart' //
-    if (dart.library.html) 'src/browser/native_interop.dart'
-    if (dart.library.io) 'src/server/native_interop.dart';
 export 'src/stub/run_app_stub.dart'
     if (dart.library.html) 'src/browser/run_app.dart'
     if (dart.library.io) 'src/server/run_app.dart' show runApp;