+++ packages/jaspr/lib/src/browser/browser_binding.dart
@@ -7,7 +7,7 @@ import 'package:meta/meta.dart';
 import '../foundation/basic_types.dart';
 import '../foundation/binding.dart';
 import '../framework/framework.dart';
-import 'dom_renderer.dart';
+import 'dom_render_object.dart';
 import 'js_data.dart';
 
 final _queryReg = RegExp(r'^(.*?)(?:\((\d+):(\d+)\))?$');
@@ -30,13 +30,13 @@ class BrowserAppBinding extends AppBinding with ComponentsBinding {
   }
 
   @override
-  Renderer createRenderer() {
+  RenderObject createRootRenderObject() {
     var attachMatch = _queryReg.firstMatch(attachTarget)!;
     var target = attachMatch.group(1)!;
     var from = int.tryParse(attachMatch.group(2) ?? '');
     var to = int.tryParse(attachMatch.group(3) ?? '');
 
-    return BrowserDomRenderer(document.querySelector(target)!, from, to);
+    return RootDomRenderObject(document.querySelector(target)!, from, to);
   }
 
   final Map<String, dynamic> _rawState = {};