+++ packages/jaspr/lib/src/server/native_interop.dart
@@ -1,5 +1,0 @@
-import '../framework/framework.dart';
-
-extension NativeDomNode on RenderElement {
-  dynamic get nativeElement => renderData;
-}